+++ scCloud/scCloud/__main__.py
@@ -1,5 +1,5 @@
 """
-Single cell RNA-Seq tools.
+Single-cell cloud-based RNA-Seq tools.
 
 Usage:
   scCloud <command> [<args>...]
@@ -10,11 +10,11 @@ Sub-commands:
   Preprocessing:
     aggregate_matrix        Aggregate cellranger-outputted channel-specific count matrices into a single count matrix. It also enables users to import metadata into the count matrix.
   Demultiplexing:
-    hashing                 Demultiplex cells/nuclei based on DNA barcodes for cell-hashing and nuclei-hashing data.
+    demuxEM                 Demultiplex cells/nuclei based on DNA barcodes for cell-hashing and nuclei-hashing data.
   Analyzing:
     cluster                 Perform first-pass analysis using the count matrix generated from 'aggregate_matrix'. This subcommand could perform low quality cell filtration, batch correction, variable gene selection, dimension reduction, diffusion map calculation, graph-based clustering, tSNE visualization. The final results will be written into h5ad-formatted file, which Seurat could load. 
     de_analysis             Detect markers for each cluster by performing differential expression analysis per cluster (within cluster vs. outside cluster). DE tests include Welch's t-test, Fisher's exact test, Mann-Whitney U test. It can also calculate AUROC values for each gene. 
-    annotate_cluster        This subcommand is used to automatically annotate cell types for each cluster based on existing markers. Currently, it only works for human and mouse immune cells.
+    annotate_cluster        This subcommand is used to automatically annotate cell types for each cluster based on existing markers. Currently, it works for human/mouse immune/brain cells.
   Plotting:
     plot                    Make static plots, which includes plotting tSNEs by cluster labels and different groups.
     iplot                   Make interactive plots using plotly. The outputs are HTML pages. You can visualize diffusion maps with this sub-command.