+++ scCloud/scCloud/commands/DemuxEM.py
@@ -1,13 +1,13 @@
 from .Base import Base
-from ..pipeline import run_hashing_pipeline
+from ..pipeline import run_demuxEM_pipeline
 
-class Hashing(Base):
+class DemuxEM(Base):
     """
-Run the hashing pipeline for cell-hashing/nuclei-hashing.
+Run the demuxEM pipeline for cell-hashing/nuclei-hashing data.
 
 Usage:
-  scCloud hashing --hash-type <type> [options] <input_adt_csv_file> <input_raw_gene_bc_matrices_h5.h5> <output_name>
-  scCloud hashing -h
+  scCloud demuxEM --hash-type <type> [options] <input_adt_csv_file> <input_raw_gene_bc_matrices_h5.h5> <output_name>
+  scCloud demuxEM -h
 
 Arguments:
   input_adt_csv_file                      Input ADT (antibody tag) count matrix in CSV format.
@@ -22,11 +22,11 @@ Options:
 
   --min-num-genes <number>                   We only demultiplex cells/nuclei with at least <number> expressed genes. [default: 100]
   --max-background-probability <prob>        Any cell/nucleus with no less than <prob> background probability will be marked as unknown. [default: 0.8]
-  --prior-on-samples <prior>                 The sparse prior put on samples. [default: 0.0]
+  --prior-on-samples <prior>                 The sparse prior put on samples.
   --random-state <seed>                      The random seed used in the KMeans algorithm to separate empty ADT droplets from others. [default: 0]
 
   --generate-diagnostic-plots                Generate a series of diagnostic plots, including the background/signal between HTO counts, estimated background probabilities, HTO distributions of cells and non-cells etc.
-  --generate-gender-plot <genes>              Generate violin plots using gender-specific genes (e.g. Xist). <gene> is a comma-separated list of gene names. 
+  --generate-gender-plot <genes>             Generate violin plots using gender-specific genes (e.g. Xist). <gene> is a comma-separated list of gene names. 
 
   -h, --help                                 Print out help information.
 
@@ -41,7 +41,7 @@ Outputs:
   output_name.gene_name.violin.png                       Optional outputs. Violin plots depicting gender-specific gene expression across samples. We can have multiple plots if a gene list is provided in '--generate-gender-plot' option.
   
 Examples:
-  scCloud hashing -p 8 --hash-type cell --generate-diagnostic-plots sample_adt.csv sample_raw_gene_bc_matrices_h5.h5 sample_output
+  scCloud demuxEM -p 8 --hash-type cell --generate-diagnostic-plots sample_adt.csv sample_raw_gene_bc_matrices_h5.h5 sample_output
     """
 
     def execute(self):
@@ -51,10 +51,10 @@ Examples:
             'genome' : self.args['--genome'],
             'min_num_genes' : int(self.args['--min-num-genes']),
             'unknown_rate' : float(self.args['--max-background-probability']),
-            'alpha' : float(self.args['--prior-on-samples']),
+            'alpha_value' : self.args['--prior-on-samples'],
             'random_state' : int(self.args['--random-state']),
             'gen_plots' : self.args['--generate-diagnostic-plots'],
             'gen_gender_plot' : self.split_string(self.args['--generate-gender-plot'])
         }
 
-        run_hashing_pipeline(self.args['<input_adt_csv_file>'], self.args['<input_raw_gene_bc_matrices_h5.h5>'], self.args['<output_name>'], **kwargs)
+        run_demuxEM_pipeline(self.args['<input_adt_csv_file>'], self.args['<input_raw_gene_bc_matrices_h5.h5>'], self.args['<output_name>'], **kwargs)