+++ scCloud/scCloud/pipeline/__init__.py
@@ -1,3 +1,3 @@
 from .pipeline import run_pipeline
-from .hashing import run_hashing_pipeline
+from .demuxEM_pipeline import run_demuxEM_pipeline
 from .cite_seq import run_cite_seq_pipeline